move cpp config into reconciliation loop (#2110)
